move database test fixture;
